+++ module/cpumod.py
